// src/types.ts
export type Movie = {
  imdbID: string;
  Title: string;
  Poster: string;
  Year: string;
  Type: string;
}

export type MovieDetails = {
  Title: string;
  Year: string;
  Runtime: string;
  Genre: string;
  Director: string;
  Language: string;
  Country: string;
  Poster: string;
  imdbRating: string;
  Plot: string;
  Actors: string;
}